rename projects files
projects.js => projects_old.js
projectstest.js => projects.js
